Add header ids in interactive playground
Fixes #228013

One of the only other places that uses marked directly for rendering
